+++ emms-player-mpv.el
@@ -112,5 +112,12 @@
   (let ((cmd (emms-player-mpv--format-command "seek %d absolute" sec)))
     (call-process-shell-command cmd nil nil nil)))
 
+;; Display deprecation warning
+(mapc (lambda (msg)
+        (display-warning '(emms emms-player-mpv) msg))
+      '("emms-player-mpv will be merged into EMMS."
+        "Please delete the emms-player-mpv package."
+        "See https://github.com/dochang/emms-player-mpv for details."))
+
 (provide 'emms-player-mpv)
 ;;; emms-player-mpv.el ends here