+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased ##
 
+### Added ###
+
+  - Add deprecation warning.
+
 ## 0.1.0 - 2018-04-09 ##
 
 ### Changed ###